# Speedup of blocks_without_reward_query