--- a/apps/explorer/priv/repo/migrations/20191208135613_block_rewards_block_hash_partial_index.exs
+++ b/apps/explorer/priv/repo/migrations/20191208135613_block_rewards_block_hash_partial_index.exs
@@ -1,0 +1,9 @@
+defmodule Explorer.Repo.Migrations.BlockRewardsBlockHashPartialIndex do
+  use Ecto.Migration
+
+  def change do
+    execute(
+      "CREATE INDEX IF NOT EXISTS block_rewards_block_hash_partial_index on block_rewards(block_hash) WHERE address_type='validator';"
+    )
+  end
+end